add FromSigner trait, eliminate dupe code
